refactor: Update header redirect in CFrontController
This commit updates the header redirect in the CFrontController class to point to the correct location. The previous redirect was set to "/UniRent/CStudent/home" but has been changed to "/UniRent/Test/test". This ensures that the user is redirected to the appropriate page when a method or controller is not found.