refactor: Improve password hashing in EStudent class
The code changes in the `EStudent` class improve the password hashing mechanism. Instead of relying on the `password_get_info` function to determine if the password is already hashed, a new `isPasswordHashed` method is introduced. This method checks if the password starts with the bcrypt hash identifier (`$2y$10$`) to determine if it is already hashed. If it is not hashed, the password is hashed using the `password_hash` function.