refactor: Add enctype attribute to registration form for image upload
The code changes add the enctype attribute to the registration form in order to support image uploads. This allows users to upload a profile picture during the registration process. The enctype attribute is set to "multipart/form-data" to handle file uploads correctly. This enhancement improves the functionality and user experience of the registration form.