refactor: Add about page functionality
The code changes in VUser.php and CUser.php add the functionality for the about page. The about method is added to the VUser class, which displays the 'about.tpl' template. The about method is also added to the CUser class, which calls the about method in the VUser class. This enhancement allows users to access and view the about page on the website.