refactor: Add contact page functionality
The code changes in VUser.php and CUser.php add the functionality for the contact page. The contact method is added to the VUser class, which displays the 'contact.tpl' template. The contact method is also added to the CUser class, which calls the contact method in the VUser class. This enhancement allows users to access and view the contact page on the website.