refactor: Add contact and about page functionality
The code changes in CStudent.php and VStudent.php add the functionality for the contact and about pages. The contact and about methods are added to the CStudent class, which calls the respective methods in the VStudent class. This enhancement allows users to access and view the contact and about pages on the website.