Refactor COwner.php and CAccommodation.php
This commit refactors the COwner.php and CAccommodation.php files. In COwner.php, the variable $disabled is now assigned the value of $accomm->getStatus(). In CAccommodation.php, two new static methods, deactivate() and activate(), have been added to handle the deactivation and reactivation of accommodations. These changes improve the functionality and maintainability of the codebase.