Refactor CStudent and publicProfileFromStudent.tpl to use consistent URLs for contracts and reviews
Update the URLs in the publicProfileFromStudent.tpl template to use the correct routes for displaying contracts and reviews from the perspective of a student. Also, refactor the CStudent class in CStudent.php to use the same URLs for retrieving contracts and reviews. This change ensures consistency in the application and improves the overall user experience.