Refactor login.tpl to add cookie consent modal
Add a cookie consent modal to the login.tpl template in the Smarty directory. The modal includes a card with a cookie SVG icon, a heading, and a description. The modal is displayed if the user's browser does not have cookies enabled. Two JavaScript functions, "on" and "off", are added to show and hide the modal, respectively.

Refactor login.tpl to add cookie consent modal