chore: Update composer.json with new namespace mapping
This commit updates the composer.json file to include a new namespace mapping for the "Installation" directory. The previous mapping only included the "Updater" directory, but now the "Installation" directory is also mapped to the "Installation\\" namespace. This change ensures that the correct namespaces are used for the respective directories in the codebase.