Refactor session handling in CAdmin.php and CFrontController.php
This commit refactors the session handling in CAdmin.php and CFrontController.php. In CAdmin.php, the code that sets the session element "id" to null when logging in as an admin is removed. In CFrontController.php, the condition for checking the session element "id" is modified to also check for the session element "userType" to be different from 'Admin' and 'undefined'. These changes improve the security and logic of the session management in the application.