Refactor Error.tpl to update ban reactivation message
This commit updates the ban reactivation message in the Error.tpl file. The previous message stated that the account would be reactivated by tomorrow, but it has been changed to indicate that the account will be reactivated in 5 days. This change provides more accurate information to the users and aligns with the actual reactivation timeline.