Refactor ban reactivation message in Error.tpl
Update the ban reactivation message in the Error.tpl file to provide more accurate information to users. The previous message stated that the account would be reactivated by tomorrow, but it has been changed to indicate that the account will be reactivated in 5 days from the request submission. This aligns with the actual reactivation timeline and improves user understanding.