Refactor addAccommodation.tpl and hiddenInputHandling.js
Update addAccommodation.tpl to include Javascript variables and files for the 'addAccommodation' page. This change ensures that the necessary scripts are loaded correctly.

In hiddenInputHandling.js, update the logic to handle different files. For 'addAccommodation', update the hidden input values for 'men', 'women', 'animals', and 'smokers'. For 'registerStudent', update the hidden input values for 'smoker' and 'animals'. These updates improve the functionality and user experience of the respective pages.